requirements.txt
py-clob-client>=0.29.0
web3>=6.0.0
python-dotenv>=1.0.0
websockets>=12.0
tenacity>=8.2.0
requests>=2.31.0
httpx>=0.27.0
